pyproject.toml
[project]
name = "office-tools"
version = "1.0.0"
description = "Office 文档批量处理工具集 - 支持 Word 和 Excel 文档的关键词替换"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Claude Code Community", email = "noreply@github.com"}
]
license = {text = "MIT"}
keywords = [
    "document",
    "replacement",
    "word",
    "excel",
    "docx",
    "xlsx",
    "batch",
    "office",
    "automation",
]

dependencies = [
    "python-docx>=1.2.0",
    "openpyxl>=3.1.5",
    "xlrd>=2.0.2",
]
